taskmanager/project_service/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
django-redis==5.3.0
